package rename scala lib
