+++ src/main/java/nl/knaw/dans/ingest/core/legacy/DepositIngestTaskFactoryWrapper.java
@@ -26,8 +26,8 @@ import nl.knaw.dans.ingest.core.config.DataverseConfigScala;
 import nl.knaw.dans.ingest.core.config.HttpServiceConfig;
 import nl.knaw.dans.ingest.core.config.IngestFlowConfig;
 import nl.knaw.dans.ingest.core.service.EventWriter;
-import nl.knaw.dans.lib.dataverse.DataverseInstance;
-import nl.knaw.dans.lib.dataverse.DataverseInstanceConfig;
+import nl.knaw.dans.lib.scaladv.DataverseInstance;
+import nl.knaw.dans.lib.scaladv.DataverseInstanceConfig;
 import scala.Option;
 import scala.collection.immutable.List;
 import scala.collection.immutable.Map;
