+++ src/main/scala/nl.knaw.dans.easy.dd2d/DatasetEditor.scala
@@ -17,10 +17,10 @@ package nl.knaw.dans.easy.dd2d
 
 import nl.knaw.dans.easy.dd2d.mapping.{ AccessRights, License }
 import nl.knaw.dans.easy.dd2d.migrationinfo.BasicFileMeta
-import nl.knaw.dans.lib.dataverse.model.dataset.Embargo
-import nl.knaw.dans.lib.dataverse.model.file.FileMeta
-import nl.knaw.dans.lib.dataverse.model.file.prestaged.PrestagedFile
-import nl.knaw.dans.lib.dataverse.{ DatasetApi, DataverseInstance }
+import nl.knaw.dans.lib.scaladv.model.dataset.Embargo
+import nl.knaw.dans.lib.scaladv.model.file.FileMeta
+import nl.knaw.dans.lib.scaladv.model.file.prestaged.PrestagedFile
+import nl.knaw.dans.lib.scaladv.{ DatasetApi, DataverseInstance }
 import nl.knaw.dans.lib.error.{ TraversableTryExtensions, TryExtensions }
 import nl.knaw.dans.lib.logging.DebugEnhancedLogging
