+++ src/main/scala/nl.knaw.dans.easy.dd2d/fieldbuilders/CompoundFieldBuilder.scala
@@ -16,7 +16,7 @@
 package nl.knaw.dans.easy.dd2d.fieldbuilders
 
 import nl.knaw.dans.easy.dd2d.mapping.JsonObject
-import nl.knaw.dans.lib.dataverse.model.dataset.{ CompoundField, MetadataField }
+import nl.knaw.dans.lib.scaladv.model.dataset.{ CompoundField, MetadataField }
 
 import scala.collection.mutable
