+++ src/main/scala/nl.knaw.dans.easy.dd2d/fieldbuilders/PrimitiveFieldBuilder.scala
@@ -15,7 +15,7 @@
  */
 package nl.knaw.dans.easy.dd2d.fieldbuilders
 
-import nl.knaw.dans.lib.dataverse.model.dataset.{ MetadataField, PrimitiveMultipleValueField, PrimitiveSingleValueField }
+import nl.knaw.dans.lib.scaladv.model.dataset.{ MetadataField, PrimitiveMultipleValueField, PrimitiveSingleValueField }
 
 import scala.collection.mutable
