+++ src/main/scala/nl.knaw.dans.easy.dd2d/migrationinfo/BasicFileMeta.scala
@@ -15,7 +15,7 @@
  */
 package nl.knaw.dans.easy.dd2d.migrationinfo
 
-import nl.knaw.dans.lib.dataverse.model.file.prestaged.PrestagedFile
+import nl.knaw.dans.lib.scaladv.model.file.prestaged.PrestagedFile
 
 case class BasicFileMeta(label: String,
                          directoryLabel: Option[String],
