+++ src/test/scala/nl.knaw.dans.easy.dd2d/DepositSpec.scala
@@ -15,7 +15,7 @@
  */
 package nl.knaw.dans.easy.dd2d
 
-import nl.knaw.dans.lib.dataverse.model.file.FileMeta
+import nl.knaw.dans.lib.scaladv.model.file.FileMeta
 
 import java.nio.file.Paths
